+++ service/supp.go
@@ -17,7 +17,7 @@ func (serv *SupplementaryService) GetRandomGoods(userId int) ([]entities.Good, e
 	return serv.repo.GetRandomGoods(userId)
 }
 
-func (serv *SupplementaryService) TransferMoney(senderId int, receiverId int, amount float64) error {
+func (serv *SupplementaryService) TransferMoney(senderId int, receiverId int, amount int) error {
 	return serv.repo.TransferMoney(senderId, receiverId, amount)
 }
